first commit 修改

+++ src/chapter02/Example01.java
@@ -3,7 +3,7 @@ package chapter02;
 public class Example01 {
 	public static void main(String[] args) {
 		int num = 4;
-		byte b = num;
+		byte b = (byte) num;
 		System.out.println(b);
 	}
 }
